chapter shader

+++ shader/include/shader.h
@@ -83,10 +83,10 @@ class Shader {
         glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value); 
     }
     void setInt(const std::string &name, int value) const {
-        glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
+        glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
     }
     void setFloat(const std::string &name, float value) const {
-        glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+        glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
     }
 
     private:
